% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{highlight_string}
\alias{highlight_string}
\title{Highlight a character vector}
\usage{
highlight_string(code, language = "r", output = "ESC_ANSI",
  theme = "solarized-light", theme_path = package_file("themes"),
  language_path = package_file("langDefs"))
}
\arguments{
\item{code}{\code{[character()]}\cr The code to highlight. Multiple lines will be joined by newlines.}

\item{language}{\code{[character(1)]} default: \sQuote{r}\cr The language,
one of \Sexpr[stage=render, results=rd]{highlite:::choices_rd(highlite:::choices(highlite:::package_file("langDefs")))} with
the default \code{language_path}.}

\item{output}{\code{[character(1)]} default: \sQuote{ESC_ANSI}\cr The output type, one of
\Sexpr[stage=render, results=rd]{highlite:::choices_rd(names(highlite:::output_types()))}.}

\item{theme}{\code{[character(1)]} default: \sQuote{solarized-light}\cr The theme, one of
\Sexpr[stage=render, results=rd]{highlite:::choices_rd(highlite:::choices(highlite:::package_file("themes")))} with the default
\code{theme_path}.}

\item{theme_path}{\code{[character(1)]}\cr The directory containing \href{http://andre-simon.de/doku/highlight/en/highlight.php}{highlight} themes.}

\item{language_path}{\code{[character(1)]}\cr The directory containing \href{http://andre-simon.de/doku/highlight/en/highlight.php}{highlight} language definitions.}
}
\description{
This function uses Andre Simon's
\href{http://andre-simon.de/doku/highlight/en/highlight.php}{highlight}
package to highlight code in
\Sexpr[stage=render]{length(highlite:::choices(highlite:::package_file("langDefs")))} programming
languages, markup languages and configuration files. Available output
formats include \sQuote{HTML}, \sQuote{ANSI escape codes}, \sQuote{TeX},
\sQuote{LaTeX} and \sQuote{SVG}.
}

